Add integration tests for various compression types
This commit is to ensure that we have basic coverage to create and
insert data into an appendonly row or column oriented tables with
quicklz, zstd, zlib, or rle. It uses get_ao_compression_ratio() function
to check insert succeess by comparing the uncompressed/compressed sizes.

It includes API testing of invalid scenarios, such as invalid compress
level or unsupported compresstype in a table format (e.g. rle is not
supported with appendonly row oriented tables).

Co-authored-by: Melanie Plageman <mplageman@pivotal.io>