Fix scan over CO table with zero projection columns
When column-oriented table scan state has 'num_proj_atts == 0', the
scan is never finished, as this case is not handled by
'open_next_scan_seg()' in aocsam.

Fix this by adding an execution branch for that case.

Due to the change, when CO table scan is over zero projected columns,
no blocks of CO table will be read.

Co-authored-by: Ivan Leskin <leskin.in@arenadata.io>