Remove VACUUM FREEZE workaround from integration tests
Upgrade performs VACUUM FREEZE and sets frozenxids on the new cluster
before restoring anything to ensure that the tables created in initdb
have relfrozenxid and relmxid values that can coexist with the old
cluser's xid counters.

Issue is when the new cluster has higher next xid than the old cluster.
In that case when we replace the next xid of the new cluster with the
old cluster's we're inducing a wrapparound.  We noticed this in the
upgrade intergration tests because a freshly initialized 6 cluster has a
higher next xid than a freshly initialized 5 cluster. This would produce
errors like "ERROR: found xmin 570 from before relfrozenxid 739".

In order to address this we can move the VACUUM FREEZE to after the
point of updating relfrozenxid on the relations.

Co-authored-by: Adam Berlin <aberlin@pivotal.io>