CPhysical now constructs a CEnfdDistribution
This commit introduces a simple change in interface where we shift the
responsibility of constructing a CEnfdDistribution object from
CReqdPropPlan object into the CPhysical object. There should be no
functional change.

This is necessary because the current separation between PdsRequired and
Edm causes us to either copy-paste code to cope with non-trivial cases,
or (when we're not lucky) unable to do our job. In a subsequent commit
we're gonna build on top of this new (and more natural) interface so we
can compute the EDistributionMatching while computing a required
distribution spec.