Add orca index only scan costing formula
Planner costs index-only-scan relative to index-scan. In the case of
index-only-scan, `RelOptInfo->allvisfrac` can be used to reduce the cost
when the visimap contains all-visible blocks. Thus, if index-only-scan
is possible, it will be favored.

With this commit, Orca also costs index-only-scan relative to index-scan
and scales the cost based on percentage of all-visible blocks in the
visimap. However, the overall cost formula remains quite different from
Planner in that it doesn't dynamically calculate index selectivity.