Orca rejects partition index if root and leaf mismatch
GPDB6 and lower assumes that the root and leaf partitions have the same
underlying column structure. That assumption can be broken when an
exchange partition with or without same dropped columns as root is
inserted into the partition table. Further complicating the matter is
that Ora always uses the root partition to construct index metadata. If
we detect a mismatch in the index and relation metadata, then we will
not consider the index.

Co-authored-by: Ekta Khanna <ekhanna@vmware.com>